ForumViewContext migrated to JPA

+++ src/main/java/com/tysanclan/site/projectewok/util/forum/ForumViewContext.java
@@ -1,17 +1,17 @@
 /**
  * Tysan Clan Website
  * Copyright (C) 2008-2013 Jeroen Steenbeeke and Ties van de Ven
- *
+ * <p>
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
- *
+ * <p>
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- *
+ * <p>
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
@@ -34,23 +34,23 @@ import javax.persistence.EntityManager;
  * @author Jeroen Steenbeeke
  */
 public interface ForumViewContext extends Serializable {
-	List<ForumCategory> getCategories(EntityManager em, User viewer, long offset,
-			long count);
+	List<ForumCategory> getCategories(EntityManager em, User viewer, int offset,
+									  int count);
 
 	int countCategories(EntityManager em, User viewer);
 
 	List<Forum> getForums(EntityManager em, ForumCategory context, User viewer,
-			long offset, long count);
+						  int offset, int count);
 
 	int countForums(EntityManager em, ForumCategory context, User viewer);
 
 	List<ForumThread> getThreads(EntityManager em, Forum context, User viewer,
-			long offset, long count);
+								 int offset, int count);
 
 	int countThreads(EntityManager em, Forum context, User viewer);
 
 	List<ForumPost> getPosts(EntityManager em, ForumThread context, User viewer,
-			long offset, long count);
+							 int offset, int count);
 
 	int countPosts(EntityManager em, ForumThread context, User viewer);
